FinalArbBot/requirements.txt
py-clob-client>=0.18.0
websocket-client>=1.6.0
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.31.0
aiohttp>=3.9.0
python-telegram-bot>=20.0
